kubernetes/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
psycopg2-binary==2.9.9
redis==5.0.3
jinja2==3.1.3
prometheus-client==0.20.0
